requirements.txt
# -------- API Layer --------
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic[email]

# -------- Database --------
sqlalchemy>=2.0
aiosqlite>=0.20

# -------- Validation --------
pydantic>=2.7
pydantic-settings>=2.3

# -------- Auth --------
python-jose[cryptography]>=3.3
bcrypt>=4.0.0
Authlib>=1.3

# -------- Forms / Uploads --------
python-multipart
python-dotenv

# -------- Networking --------
httpx>=0.27
requests>=2.32
Pillow
itsdangerous

# -------- LangChain Stack --------
langchain>=0.2
langgraph>=0.2
google-genai
# -------- Vector / DB --------
supabase>=2.4
pypdf
langchain-community
langchain-tavily
tavily-python
langchain-groq
google-auth
pinecone
livekit-agents
livekit-plugins-google